requirements.txt
Flask==2.2.5
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
email-validator==2.0.0
Werkzeug==2.2.2
